+++ src/gui/components/theme.py
@@ -1,0 +1,1 @@
+styles = {}
