+++ src/gui/screens/__init__.py
@@ -1,4 +1,9 @@
-from .base import Screen, MenuScreen, Alert, Prompt, QRAlert, ProgressScreen
+from .screen import Screen
+from .menu import Menu
+from .alert import Alert
+from .prompt import Prompt
+from .qralert import QRAlert
+from .progress import Progress
 from .input import PinScreen, InputScreen, DerivationScreen
 from .mnemonic import MnemonicScreen, NewMnemonicScreen, RecoverMnemonicScreen
 from .bitcoin import XPubScreen, WalletScreen
