+++ src/keystore/core.py
@@ -1,3 +1,5 @@
+"""Base classes for inheritance"""
+
 class KeyStoreError(Exception):
     pass
 
@@ -5,8 +7,4 @@ class PinError(KeyStoreError):
 	pass
 
 class KeyStore:
-    STATUS_SETUP    = 2 # PIN is not set
-    STATUS_LOCKED   = 3 # PIN is set but not entered
-    STATUS_UNLOCKED = 4 # normal operation
-    STATUS_BLOCKED  = 5 # no PIN attempts left - permanently blocked
-    STATUS_ERROR    = 6 # something bad happened
+    pass
