+++ src/main.py
@@ -39,5 +39,4 @@ def main():
     specter.start()
 
 if __name__ == '__main__':
-    # main()
-    pass
+    main()
